+++ frontend/src/pages/organization/SettingsPage/components/OrgWorkflowIntegrationTab/index.ts
@@ -1,0 +1,1 @@
+export * from "./OrgWorkflowIntegrationTab";
