--- frontend/src/pages/organization/SsoPage/components/SsoTabGroup/index.tsx
@@ -1,1 +1,0 @@
-export { SsoTabGroup } from "./SsoTabGroup";
